Refactor sensitive field handling and move logger utilities
