+++ agave/fastapi/middlewares/__init__.py
@@ -1,13 +1,7 @@
 from .error_handlers import AgaveErrorHandler
-from .request_logger import (
-    SENSITIVE_REQUEST_MODEL_FIELDS,
-    SENSITIVE_RESPONSE_MODEL_FIELDS,
-    AgaveRequestLogger,
-)
+from .request_logger import AgaveRequestLogger
 
 __all__ = [
     'AgaveErrorHandler',
     'AgaveRequestLogger',
-    'SENSITIVE_RESPONSE_MODEL_FIELDS',
-    'SENSITIVE_REQUEST_MODEL_FIELDS',
 ]
